feat: migrate app to ts
