+++ .eslintrc.json
@@ -1,6 +1,6 @@
 {
-  "root": true,
-  "env": {
+    "root": true,
+    "env": {
     "es2020": true,
     "jasmine": true,
     "jest": true,
@@ -20,13 +20,16 @@
     "eslint:recommended",
     "plugin:node/recommended",
     "airbnb-base",
-    "prettier"
+    "prettier",
+    "plugin:@typescript-eslint/recommended"
   ],
   "rules": {
     "import/extensions": "off",
     "linebreak-style": "off",
     "node/no-unsupported-features/es-syntax": "off",
     "no-underscore-dangle": "off",
-    "import/prefer-default-export": "off"
-  }
+    "import/prefer-default-export": "off",
+  },
+  "parser": "@typescript-eslint/parser",
+  "plugins": ["@typescript-eslint"]
 }
