+++ src/resources/boards/board.router.ts
@@ -43,4 +43,4 @@ router.route('/:id').delete(async (req, res) => {
   }
 });
 
-module.exports = router;
+export { router };
