+++ src/resources/tasks/task.router.ts
@@ -41,4 +41,4 @@ router.route('/:boardId/tasks/:taskId').delete(async (req, res) => {
   }
 });
 
-module.exports = router;
+export { router };
