+++ src/resources/users/user.router.ts
@@ -44,4 +44,4 @@ router.route('/:id').delete(async (req, res) => {
   }
 });
 
-module.exports = router;
+export { router };
